tests: ui: compile-fail: pin_data: no_pin_on_phantompinned: fix warning
The `#[pin_data]` macro checks for fields with the type `PhantomPinned`
and then emits an error if it's not marked with `#[pin]`. This check
replaces the originally typed path with `::core::marker::PhantomPinned`
and thus any imports of the `PhantomPinned` stuff are not used. Opened
an issue: https://github.com/Rust-for-Linux/pin-init/issues/51

Signed-off-by: Benno Lossin <benno.lossin@proton.me>