add `raw_ref_op` feature gate to examples & tests
Due to `raw_ref_op` not being stable until rustc 1.82 (which is the
"proper" msrv, when no default features), we have to explicitly enable
it using a feature gate.

This commit also adds a flag `RUSTC_RAW_REF_OP_IS_STABLE` to the
`build.rs` to communicate with cargo for conditional compilation
as to whether the `raw_ref_op` feature gate is needed, based on the
RUSTC version.

Finally it adds the feature gate behind the conditional flag mentioned
above to the examples and tests, so they don't fail the nightly-msrv
checks (where `raw_ref_op` is currently unstable).

Signed-off-by: Antonio Hickey <contact@antoniohickey.com>